// interfaces/artistResponse.ts
// Generated by https://quicktype.io

export interface Response {
  ok: boolean;
  data: Artist[];
}

export interface Artist {
  _id: string;
  image: string;
  followers: number;
  name: string;
  popularity: number;
  genres: string[];
  tracks: Track[];
  albums: Album[];
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export interface Album {
  _id: string;
  image: string;
  title: string;
  releaseDate: string;
  totalTracks: number;
  tracks: string[];
  artist: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}

export interface Track {
  _id: string;
  title: string;
  duration: number;
  trackNumber: number;
  trackAudio: string;
  album: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}